Move edit projection components to higher directory
