+++ src/components/projections/Edit/shared.ts
@@ -1,0 +1,1 @@
+export const TITLE_ID = 'edit-projection-dialog-title';
